+++ frontend/src/components/molecules/messages/Message.tsx
@@ -16,6 +16,7 @@ import { useLayoutMaxWidth } from 'hooks/useLayoutMaxWidth';
 
 import type { IAction, IMessageElement, IStep } from 'client-types/';
 
+import BlinkingCursor from '../BlinkingCursor';
 import { Messages } from './Messages';
 
 interface Props {
@@ -62,11 +63,13 @@ const Message = memo(
     }
 
     const isAsk = message.waitForAnswer;
+    const isUserMessage = message.type === 'user_message';
 
     return (
       <Box
         sx={{
-          color: 'text.primary'
+          color: 'text.primary',
+          position: 'relative'
         }}
         className="step"
       >
@@ -96,31 +99,81 @@ const Message = memo(
               overflowX: 'auto'
             }}
           >
-            <Author message={message} show={showAvatar}>
-              <Stack alignItems="flex-start" minWidth={150}>
-                <MessageContent
-                  elements={elements}
-                  message={message}
-                  preserveSize={!!message.streaming || !defaultCollapseContent}
-                  allowHtml={allowHtml}
-                  latex={latex}
-                />
+            {isUserMessage ? (
+              <Box display="flex" flexDirection="column" width="100%">
+                <Box
+                  sx={{
+                    px: 2.5,
+                    py: 1,
+                    borderRadius: '1.5rem',
+                    backgroundColor: 'background.paper',
+                    maxWidth: '90%',
+                    ml: 'auto'
+                  }}
+                >
+                  <MessageContent
+                    elements={elements}
+                    message={message}
+                    preserveSize={
+                      !!message.streaming || !defaultCollapseContent
+                    }
+                    allowHtml={allowHtml}
+                    latex={latex}
+                  />
+                </Box>
                 <DetailsButton
                   message={message}
                   opened={showDetails}
                   onClick={() => setShowDetails(!showDetails)}
                   loading={isRunning && isLast}
                 />
-                {!isRunning && isLast && isAsk && (
-                  <AskUploadButton onError={onError} />
-                )}
-                {actions?.length ? (
-                  <MessageActions message={message} actions={actions} />
-                ) : null}
-                <MessageButtons message={message} />
-              </Stack>
-            </Author>
+              </Box>
+            ) : (
+              <Author message={message} show={showAvatar}>
+                <Stack alignItems="flex-start" minWidth={150}>
+                  <MessageContent
+                    elements={elements}
+                    message={message}
+                    preserveSize={
+                      !!message.streaming || !defaultCollapseContent
+                    }
+                    allowHtml={allowHtml}
+                    latex={latex}
+                  />
+                  <DetailsButton
+                    message={message}
+                    opened={showDetails}
+                    onClick={() => setShowDetails(!showDetails)}
+                    loading={isRunning && isLast}
+                  />
+                  {!isRunning && isLast && isAsk && (
+                    <AskUploadButton onError={onError} />
+                  )}
+                  {actions?.length ? (
+                    <MessageActions message={message} actions={actions} />
+                  ) : null}
+                  <MessageButtons message={message} />
+                </Stack>
+              </Author>
+            )}
           </Stack>
+          {isLast && isRunning && !message.streaming && (
+            <Box
+              sx={{
+                position: 'absolute',
+                bottom: -5,
+                left: 38,
+                boxSizing: 'border-box',
+                mx: 'auto',
+                maxWidth: layoutMaxWidth,
+                px: 2,
+                display: 'flex',
+                flexDirection: 'column'
+              }}
+            >
+              <BlinkingCursor />
+            </Box>
+          )}
         </Box>
         {message.steps && showDetails && (
           <Messages