+++ frontend/src/components/organisms/chat/inputBox/UploadButton.tsx
@@ -1,12 +1,13 @@
 import { useUpload } from 'hooks';
 import { useRecoilValue } from 'recoil';
 
-import AttachFile from '@mui/icons-material/AttachFile';
 import { IconButton, Theme, Tooltip, useMediaQuery } from '@mui/material';
 
 import { FileSpec } from '@chainlit/react-client';
 
 import { Translator } from 'components/i18n';
+
+import AttachmentIcon from 'assets/attachment';
 
 import { projectSettingsState } from 'state/project';
 
@@ -55,7 +56,7 @@ const UploadButton = ({
           size={size}
           {...getRootProps({ className: 'dropzone' })}
         >
-          <AttachFile fontSize={size} />
+          <AttachmentIcon fontSize={size} />
         </IconButton>
       </span>
     </Tooltip>